moving gateway server into subdir
