+++ gateway/go.mod
@@ -2,8 +2,9 @@ module porters
 
 go 1.20
 
+require github.com/redis/go-redis/v9 v9.4.0
+
 require (
 	github.com/cespare/xxhash/v2 v2.2.0 // indirect
 	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/redis/go-redis/v9 v9.4.0 // indirect
 )
