+++ gateway/proxy/proxy.go
@@ -1,4 +1,4 @@
-package gateway
+package proxy
 
 import (
     "log"
